logica do carrinho e itens funcionando

+++ order/models.py
@@ -4,4 +4,4 @@ from django.contrib.auth.models import User
 class Order(models.Model):
     client = models.ForeignKey(User, on_delete=models.CASCADE)
     date = models.DateTimeField(auto_now_add=True)
-    total = models.DecimalField(max_digits=10, decimal_places=2)
+    total = models.FloatField()
